Remove need for noaa oisst reference file in tests

+++ fixtures/generate_test_store.py
@@ -14,10 +14,10 @@ lat = np.linspace(-90, 90, lat_dim)
 lon = np.linspace(-180, 180, lon_dim, endpoint=False)
 
 # Initialize variables with random data
-CDD0 = xr.DataArray(np.random.rand(time_dim, lat_dim, lon_dim), dims=("time", "lat", "lon"), name="CDD0")
-DISPH = xr.DataArray(np.random.rand(time_dim, lat_dim, lon_dim), dims=("time", "lat", "lon"), name="DISPH")
-FROST_DAYS = xr.DataArray(np.random.rand(time_dim, lat_dim, lon_dim), dims=("time", "lat", "lon"), name="FROST_DAYS")
-GWETPROF = xr.DataArray(np.random.rand(time_dim, lat_dim, lon_dim), dims=("time", "lat", "lon"), name="GWETPROF")
+CDD0 = xr.DataArray(np.random.rand(time_dim, lat_dim, lon_dim).astype(np.uint8), dims=("time", "lat", "lon"), name="CDD0")
+DISPH = xr.DataArray(np.random.rand(time_dim, lat_dim, lon_dim).astype(np.uint8), dims=("time", "lat", "lon"), name="DISPH")
+FROST_DAYS = xr.DataArray(np.random.rand(time_dim, lat_dim, lon_dim).astype(np.uint8), dims=("time", "lat", "lon"), name="FROST_DAYS")
+GWETPROF = xr.DataArray(np.random.rand(time_dim, lat_dim, lon_dim).astype(np.uint8), dims=("time", "lat", "lon"), name="GWETPROF")
 
 # Create dataset
 ds = xr.Dataset(
